backend/requirements.txt
fastapi
uvicorn
openai
requests
beautifulsoup4
python-dotenv
